+++ zebra-consensus/src/block.rs
@@ -245,6 +245,8 @@ where
                 .map_err(VerifyBlockError::Time)?;
             let coinbase_tx = check::coinbase_is_first(&block)?;
 
+            check::shielded_action_limits_are_valid(&block.transactions, height, &network)?;
+
             let expected_block_subsidy =
                 zebra_chain::parameters::subsidy::block_subsidy(height, &network)?;