+++ zebra-consensus/src/transaction.rs
@@ -405,6 +405,11 @@ where
             check::has_inputs_and_outputs(&tx)?;
             check::has_enough_orchard_flags(&tx)?;
             check::consensus_branch_id(&tx, req.height(), &network)?;
+            crate::block::check::shielded_action_limits_are_valid(
+                std::iter::once(&tx),
+                req.height(),
+                &network,
+            )?;
 
             // Validate the coinbase input consensus rules
             if req.is_mempool() && tx.is_coinbase() {